+++ src/ImageProcessorCore/Samplers/Resamplers/SplineResampler.cs
@@ -12,13 +12,13 @@ namespace ImageProcessorCore
     public class SplineResampler : IResampler
     {
         /// <inheritdoc/>
-        public float Radius => 2;
+        public double Radius => 2;
 
         /// <inheritdoc/>
-        public float GetValue(float x)
+        public double GetValue(double x)
         {
-            const float B = 1;
-            const float C = 0;
+            const double B = 1;
+            const double C = 0;
 
             return ImageMaths.GetBcValue(x, B, C);
         }
